spam_email_detector/requirements.txt
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
nltk==3.9.1
joblib==1.4.2
beautifulsoup4==4.12.3
requests==2.32.3
# Gmail API
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
# Optional Web UI
streamlit==1.39.0
